+++ src/test/suite/sortOrder/sortObjects/sortObjectArray.test.ts
@@ -3,7 +3,7 @@ import {SortOrder} from '../../../../sortOrder/SortOrder';
 import {ALL, JOHN, ROBERT, JOHN_PAUL, JIMMY} from './lz';
 import chai = require('chai');
 import {sortObjectArray} from '../../../../sortOrder/sortObjects/sortObjectArray';
-import {SortConfiguration} from '../../../../sortOrder';
+import {SortConfiguration} from '../../../../extensionConfiguration';
 const expect = chai.expect;
 
 const SORT_CONFIG_ASCENDING: SortConfiguration = {order: SortOrder.ascending, collator: new Intl.Collator()};